Remove instruction replacing with note
I had suggested in a previous commit to add instruction of docker-compose up, but I realize the github scripts command script/server builds the docker container and script/setup inits the database just like the instructions say. I added extra instruction to make sure docker is installed as the desktop app. I put a link with more info on installing docker in case the user is on a different system like Linux.